Fix #1: Define `height` of `<body>`
Setting `height` of `html` and `min-height` of
`body` to `100%` fixed the bug.

**Explanation**:
Although it looked like a bug, Firefox behaved
correctly.

If elements have no defined `height`, they get
the sum of the `height` of all child elements
which are in the _elements flow_.
In this case, `.toolbox` is the only _direct_
child of `<body>`.

As soon as you drag `.toolbox`, the
`position`-property changes from `static` to
`absolute`, which takes the element out of the
_elements flow_, and the `<body>` has a `height`
of `0`.

Because the `<body>` is defined as the
_containment_ of the dragged element on
[line #2 of main.js][script], it could only be
moved inside the bounds of `<body>`, which had
a `height` of `0`.

  [script]:
  https://github.com/andreas-marschke/mockup-dashboard-designer/blob/master/js/main.js#L2